+++ CHANGELOG.md
@@ -17,7 +17,9 @@ CHANGELOG for FlatCAM beta
 - minor changed in camlib.py regarding the self.check_zcut() method setting the self.z_cut to 'fail'
 - upgraded the `mirror` Tcl command to work on a selection of objects and also fixed a plot issue
 - upgraded the `skew` Tcl command to work on a selection of objects and added the ability to skew not only in degrees but also by distances
-- 
+- fixed some issues in `panelize` and `cncjob` Tcl commands
+- reversed the mirroring axis in the `mirror` Tcl command to be consistent with the action in GUI
+
 31.02.2022
 
 - fixed all the `laser` preprocessors to work correctly and the resulting GCode to be plotted